feat: update getDocument to handle md files
Also, create saveFileBlob utility function
